+++ relay-server/src/envelope.rs
@@ -106,8 +106,8 @@ pub enum ItemType {
     ProfilingSession,
     /// Profiling data
     ProfilingTrace,
-    /// Stacktrace event payload encoded in JSON
-    Stacktrace,
+    /// Profile event payload encoded in JSON
+    Profile,
 }
 
 impl ItemType {
@@ -141,7 +141,7 @@ impl fmt::Display for ItemType {
             Self::ClientReport => write!(f, "client report"),
             Self::ProfilingSession => write!(f, "profiling session"),
             Self::ProfilingTrace => write!(f, "profiling trace"),
-            Self::Stacktrace => write!(f, "stacktrace"),
+            Self::Profile => write!(f, "profile"),
         }
     }
 }
@@ -588,7 +588,7 @@ impl Item {
             | ItemType::ClientReport
             | ItemType::ProfilingSession
             | ItemType::ProfilingTrace
-            | ItemType::Stacktrace => false,
+            | ItemType::Profile => false,
         }
     }
 
@@ -612,7 +612,7 @@ impl Item {
             ItemType::ClientReport => false,
             ItemType::ProfilingSession => false,
             ItemType::ProfilingTrace => false,
-            ItemType::Stacktrace => false,
+            ItemType::Profile => false,
         }
     }
 }